Add JUnit tests for Task, ToDo, Event and Deadline
There is currently no test to check that the implementation of toString
and toTxt work correctly for these classes. If the implementation of
either malfunctions due to future commits, it could cause a cascade of
errors. For example, the tasklist.txt would be interpreted as being corrupted because the format for parsing will be wrong.

Let's write some JUnit tests for these classes to ensure that the
conversion to String and txt representation is correct.