Extract out the logic of checking Gopher message type to Message.java
Currently, Gopher's response will be styled as red background if the
response is a warning message that contains text "Please try again".
Even though such implementation works in most cases, it will introduce
styling error in the system if user creates a task with "Please try
again" as part of its name, hence affecting user experience. Also, such
way of recognizing message type and styling is not very extensible and
flexible, hence a better idea would be extract out the logic to another
class to explicitly handle this, so as to provide more flexibility when
similar features need to be developed in the future.

As a result, Message class is created to wrap around the text returned
by Gopher so that other information such as message type can be input
and pass around the program for checks. This offers more flexibility as
compared to the initial approach and hence should be included in the
latest update