Improve code quality
Parser class is too long and handles the MeowMeow output too.
The parse method is too long and violates SLAP.
The Ui class is does not handle the MeowMeow output.
There are minor SLAP violations in other classes as well.

Refactoring the Parser class to only handle parsing while the Ui class
handles the MeowMeow output allows us to practice better OOP.
Fixing SLAP violations allows us to ensure good code quality so that
the code is readable.

Let's,
* remove all code handling the MeowMeow output from the Parser class
and put it into the Ui class.
* split up long methods into shorter methods
* correct poor code quality

This allows for our code to be more readable.