Add assertions to Storage and MainWindow classes
Storage class deals with loading to or saving from new or existing save
files, hence makes assumptions about whether the save file exists or
not or whether it is empty or has valid tasks in it.

MainWindow class requires non-null instances of MeowMeow and Parser.
It also takes in user input and takes in responses from MeowMeow.
It makes assumptions that these are not null.

These assumptions could indicate bugs in the code if violated.

Let's add assertions to the Storage and MainWindow classes.

This allows us to detect assertion failure and terminate the
execution if there is a possible bug.