Shortened messages for multiple task operation command
Currently when multiple tasks are manipulated by one single command,
such as mark, unmark and update, the title of the message will be printed
for each task, which makes the whole message looked very messy and hard
to read.

Modify the logic of creating messages for those commands such that the
title only appears for once per operation instead of per task.