Main Class: add assertions to ensure FXML and anchor pane load successfully
The start method of Main class does not have checks
to confirm that FXML and AnchorPane are successfully loaded.
Checks are needed to spot potential issues during early stages
of execution of the app.

Let's add assertions to verify that FXMLLoader and AnchorPane is loaded
correctly.

These checks can help to debug and prevent further errors during the
loading of the UI.