Add archive feature
There is no way to keep old tasks but not have them in the task list.

Users may want to keep old tasks but not see it on the main task list.

Add archive ArrayList field within TaskList class which can be used to
place archive tasks in.

Having the archive be accessible through the TaskList allows commands
to have access to the archive.