Update Gui
Currently, all responses by Colress have the same
background.

This makes the interface rather bland and the commands
carried out are not fleshed out as much.

By adding different coloured backgrounds that correspond
to certain command types, it makes the interface more
pleasing to look at as a whole, and the colour coding
makes the commands and error messages stand out
more.