Check asserts created and their messages
There is currently no asserts in the code for Elster.

This means that we do not verify our assumptions about the program state
at runtime and can cause undesirable details to be exposed to the client
from error messages.

Let's add asserts into the code so that we can verify that key
assumptions stay true as well as prevent exposing undesirable details to
our clients.

This can also ensure that regression testing can be easier done when
modifications are made in the future.