Extract methods in Storage class
Storage class have repeated if else structures. Additionally, there
were long methods that had multiple levels of abstraction.

The repeated behaviours caused code duplication, whereas the violation
of SLAP made the code difficult to read for someone new to the codebase.

Let's extracts behaviours common to the methods in the Storage class to
address code duplication. Let's also extract methods such that SLAP is
obeyed by the code.

This ensures that the code is more readable to anyone that comes accross
the codebase.