Fix the issue of valid duration being treated as invalid
Currently, if user updates an event task with new valid start and end
date, but the new start date is after the current end date, Gopher will
complain that the user has given an invalid date duration, which is not
correct. This is because Gopher only checks whether the start date is
before the current end date or not, and not checking whether it is
before the new end date.

Hence, extra logic is added to determine the new start and end date
before the actual comparison to make sure that Gopher will always
analyze the given duration correctly.