Fix checkstyle mistakes
Reorder import statements and remove whitespaces at the end of the lines

This is to be in line with the coding standard. However, I did ignore the warning which involves leaving empty lines.

There is no requirement within the coding standard to pack everything together, which makes it hard to tell when one method starts and another method end. It is also in line with Oracle's coding standard.

Refer to https://www.oracle.com/java/technologies/javase/codeconventions-whitespace.html