EventTask's start and end times are now represented as LocalDateTime objects instead of Strings.
Previously, the start and end times were handled as strings, which required manual saving of the duration of the event task. After using LocalDateTime, the class can automatically parse the input duration into dates and times and this improves the efficiency of the bot.

*Updated EventTask constructor to parse date-time strings into LocalDateTime.
*Added dateTimeFormatter for consistent formatting of time and date throughout the application.
*Modified printTask and printFileFormat to suit the new data types of the start and end time of Event Tasks.

These changes improve the robustness of date-time handling and improves the efficiency of task management.