fix: Add DudeInvalidNameException.java
Users can add task with description that contains characters
that leads to data corruption.

The occurence of data corruption diminish user experience and
reduce the usefulness of the application.

Let's add DudeInvalidNameException to deal with invalid task
name.

Using exception to deal with this issue can prevents data
corruption and inform the user about it.