Fix bug with ToDo::toFileString() encoding wrong format
ToDo::toFileString() encodes the UI Task representation to a save file
instead of the intended file format by calling super.toString() instead
of toFileString().

This causes an error in decoding tasks at the booting of HypeBot by the
TasklistDecoder as Tasks that should be saved and decodeable become
encoded in a wrong format, not decodeable, then not saveable.

Change super.toString() to super.toFileString().

This will allow proper encoding and decoding.