Edit build.gradle to allow usage of shadowjar plugin
To integrate the usage of the shadowjar plugin to create a fat JAR file,
the build.gradle's dependencies had to be modified.

Specifically added the lines mainClassName = 'wenjiebot.WenJie' so that
the plugin knows where the entry point of the programme is, to ensure
the plugin runs properly.

It was done this way as given in the instructions on the se-edu website.