Refactor code in Storage.java
To keep in line with the coding standard of SLAP, and to decrease the
number of lines of code within the methods to less than 30.

By using IntelliJ's refactoring feature, I abstracted out some parts of
each method to its own new method to follow the code quality.

Let's try to ensure that we use more guard clauses.