Add support for managing notes
There is currently no support for adding and showing notes for the
respective tasks.

Users may want to record small snippets of textual information relevant
to the task that they want to keep track of.

Adding support for managing such notes could be done through the use of
an optional flag in the respective task command to add the notes.
Another command could also be added to show the details of the specific
task and the note associated with it.

Adding another flag within the task creation commands will allow users
to easily add notes to that task. Note should only be displayed when
specified by the user to reduce visual clutter.