Refactor multiple classes
There are multiple classes that contain methods which throw
invalid task number exceptions. However, the methods used to deal with
such exceptions differ throughout the different classes, and there are
instances of using try catch blocks with this exception to control the
flow of logic. The error message for InvalidTaskNumberException is also
specialised to numbers that are non-positive or larger than the
number of tasks in the task list.

Refactoring the relavant classes and removing redundant try catch
blocks allows for better readability and hence easier debugging.
Generalising the error message for InvalidTaskNumberException also
helps with providing error messages that are more relevant and easier
understand.