Improve GUI
The GUI is basic and not aesthetically pleasing.

A better GUI is needed for a better user experience.

Add a background color and speech bubbles through a css file.
Anchor UI elements and implement automatic resizing of UI elements
to allow proper window resizing.

Using CSS allows the colors and other visual properties
to be changed easily in the future.