Tweak parser to give user more accurate feedback upon invalid input
Previously program would feedback the same error message regardless of whether the command was unrecognized or if the task description was left as empty. This is not user friendly as users will be confused about what the error in their input was.

This modified code will now more accurately feedback to the user what the error in their input was so that they are better able to correct their input.