Fix Storage Write and Delete Command
Currently, the delete command will throw an uncaught exception if the
user enters a index that is not in the task array. In addition,
Storage class will create two text file instead of one text file.

Fix the bug on the Storage class which will now create just one text
file. The delete command only has a index checker to catch any
invalid index.

This ensures that the application is bug free for the user.