Add message for empty list and resolve code style
Ned shows an empty message body if the list of tasks is empty.
In addition there are several code style issues like a lack of
standardisation of where the exit message is derived from and formatting
issues in the docs/README.md file.

- When `list` command is run and if the size of the listOfTasks == 0,
then a special message is shown to the user to indicate the list is
empty.
- The exit message has been refactored to a static constant belonging to Ned class,
with a public getter method for future access to it.
- Broken links in the README.md file have been removed, and formatting
  issues resolved.