Refactor input handling and UI with FXML
Transitioned from Java code-based UI to FXML to enhance
the clarity and maintainability of the user interface.

The previous implementation used Java for constructing the UI,
which resulted in complex and hard-to-manage code. By adopting
FXML, the UI layout is now more straightforward and can be
easily modified without altering the underlying logic.

Additionally, I implemented more specific exception handling
for user input errors, particularly for incorrect date formats
and invalid command structures. This ensures that users receive
clear feedback when their input does not conform to the expected
formats, improving the overall user experience and robustness
of the application.

For example, input like `deadline homework by 2024-02-22 1900`
will now prompt the user with a specific error message,
indicating that the format is incorrect.

These enhancements contribute to cleaner code and a more
user-friendly interface, setting the stage for future
improvements and features.