Implement duplicate task detection
Luke is not able to detect duplicate tasks; two tasks of the same name can be added to the task list.

Assuming a duplicate task was added by accident, the user will not be notified and the mistake will likely go undetected.

To resolve this issue, let's implement duplicate task detection in TaskList. Any two tasks will be considered duplicates of one another if (1) their names AND types are the same OR (2) they reference the same location in memory. Additionally, Deadline tasks are now formatted to consistently store dates and times in a uniform format, regardless of the format used in the input.

Duplicate task detection should be implemented in TaskList and not other parts of the Luke codebase. This will ensure we can catch duplicate tasks as soon as Luke attempts to them to the TaskList.