Fix Statistics bug and Add User Guide
Statistics calculate tasks wrongly and there is no user guide. This
makes the statistics result inaccurate. Users cannot operate the
application freely without the user guide.

Let's refactor the Statistics addition of tasks to a single method
and marked or unmarked tasks to another method to make it easier
to read and modify in the future. We also add user guide based on the
commands that the users can input.