Add information into README regarding features
Chatbot features can be difficult to navigate for first time users.
Having a README can help one explore the features of the chatbot, to
utilise it more effectively.

Detailed descriptions of the features, as well as example usages
are reflected to make the chatbot more easy to understand and use.