Add condition for EventTask start time
The start date and time attribute for EventTask accepts any value, even
if it is a date and time that is after the end date and time.
Realistically, it is illogical.

To ensure that users add Event Task that makes sense, let's
* Add a condition that checks if the start date and time of the
  EventTask is before the end date and time
* Throw a DateTimeException if it is found that the start date and time
  of the EventTask is after the end date and time

This will inform users that the event task they add should be rectified.